requirements.txt
Flask>=2.0.3
authlib>=1.0
google-generativeai
joblib
pdfminer.six
Pillow
python-dotenv>=0.19.2
pytesseract
requests>=2.27.1
scikit-learn
